kernel: 4.19.0 CONFIG_SIGEXIT
StarlingX: Death of Arbitrary Process Notification

This exposes a new feature which may be called to request
notification when an arbitrary process changes state. The
caller specifies a pid, signal number, and event mask, and
when that pid dies, or is stopped, or anything else that
would normally cause a SIGCHLD, the kernel will send the
specified signal to the caller if the event is in the event
mask originally passed down. The siginfo_t struct will
contain the same information as would be included with SIGCHLD.

This is exposed to userspace via the prctl() call with the
PR_DO_NOTIFY_TASK_STATE option.

https://github.com/xe1gyq/linux/commit/735410f968473ca710f1e026a10186a2108f5035

No LSB modules are available.
Distributor ID: Ubuntu
Description:    Ubuntu 16.04.5 LTS
Release:        16.04
Codename:       xenial

Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>